requirements.txt
chromadb==1.0.20
openai>=1.35.0
streamlit>=1.26.0
python-dotenv>=1.0.1
pydub>=0.25.1
pillow>=10.0.0
scikit-learn>=1.3.0